328959916 - Initial draft of final beta feature set
