+++ moncli/entities/__init__.py
@@ -1,0 +1,1 @@
+from .client import MondayClient
